massive reorganization - decider and repository as base crate modules
