+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod decider;
+pub mod repository;
 
 #[cfg(test)]
-mod test;
+mod test_helpers;
