+++ src/repository/esdb/error.rs
