+++ src/repository/in_memory/versioned_with_streams/error.rs
